Reorganise Court data adaptor library errors
Moved to a more generic errors module for
storing specific errors to make it more
extensible.
