+++ lib/court_data_adaptor.rb
@@ -3,8 +3,8 @@
 require_relative 'court_data_adaptor/configuration'
 require_relative 'court_data_adaptor/configurable'
 require_relative 'court_data_adaptor/client'
+require_relative 'court_data_adaptor/errors'
 require_relative 'court_data_adaptor/resource/base'
-require_relative 'court_data_adaptor/resource/error'
 require_relative 'court_data_adaptor/resource/queryable'
 require_relative 'court_data_adaptor/resource/prosecution_case'
 require_relative 'court_data_adaptor/resource/defendant'
